+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Complete rewrite
 * Remove DOM-style rendering from the server
+* Informer class is now made using Function.inherits and its methods are set
+  using the #setMethod method
 
 ## 0.1.1 (2014-06-10)